Move the new tests to their own file
