+++ src/tests.rs
@@ -1,1 +1,2 @@
 mod aeap;
+mod verified_chats;
